Add files via upload
This project aims to build an image classification model capable of distinguishing between images of cats and dogs. The model uses a Convolutional Neural Network (CNN), which is a deep learning architecture specifically designed for processing images. The model is trained on a labeled dataset containing images of cats and dogs, where each image is labeled as either "cat" or "dog."